app/requirement.txt
openai
azure-functions
azure-identity
azure-keyvault-secrets
llama-index
llama-cpp-python
